+++ src/main/java/org/example/TransactionHelper.java
@@ -1,0 +1,35 @@
+package org.example;
+
+import org.hibernate.SessionFactory;
+import org.hibernate.Transaction;
+import org.hibernate.resource.transaction.spi.TransactionStatus;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.util.function.Supplier;
+
+@Component
+public class TransactionHelper {
+    @Autowired
+    SessionFactory sessionFactory;
+
+    public<T> T executeInTransaction(Supplier<T> action) {
+
+        var session = sessionFactory.getCurrentSession();
+        Transaction transaction = session.getTransaction();
+        if (!transaction.getStatus().equals(TransactionStatus.NOT_ACTIVE)) {
+            return action.get();
+        }
+        try {
+            session.beginTransaction();
+            T returnValue = action.get();
+            transaction.commit();
+            return returnValue;
+        } catch (Exception e) {
+            transaction.rollback();
+            throw e;
+        } finally {
+            session.close();
+        }
+    }
+}
